smart-agent-hub/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-agent-hub"
version = "0.1.0"
description = "MCP Client Agent for RAG-MCP-SERVER integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "httpx>=0.24",
    "aiohttp>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1",
]
mcp = [
    "mcp>=1.0",
]
dashboard = [
    "streamlit>=1.28",
]
all = [
    "smart-agent-hub[dev,mcp,dashboard]",
]

[project.scripts]
agent = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]